requirements.txt
flask==2.0.1
pandas==1.0.5
numpy==1.19.1
nltk==3.5
scipy==1.4.1
pickle-mixin
keras==2.4.0
tensorflow==2.2.0
spacy==2.3.2
h5py==2.10.0
requests
scikit-learn==0.23.1
google-cloud-storage